edit: extern 메서드 숨김, Exception 정의

+++ Packages/com.seonghwan.windows.interceptor/Runtime/Windows/Util/IKeyCodeHandler.cs
@@ -12,7 +12,7 @@
             // alphabets
             if (VirtualKey >= 0x41 && VirtualKey <= 0x5A)
             {
-                return (int)(32 + VirtualKey);
+                return 32 + VirtualKey;
             }
             
             // numbers
